refactor: split indexer event handlers
